Latest Code
Been working on making the thread syncs better, or actually work.
Setup the project to use GLEW now so I can do some of the OGL tutorials.
That way we I can maybe get the window to correctly render the frame as
we trace it.

220 Jobs per second to 4.5k jobs per second. Decoupling the allocation
of jobs from the main thread was a good idea.